pyproject.toml
[project]
name = "superseded"
version = "0.1.0"
description = "Local-first agentic pipeline tool — write tickets, delegate the rest"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "aiosqlite>=0.21.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.10.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[project.scripts]
superseded = "superseded.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/superseded"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in argument defaults (FastAPI Depends pattern)
    "TC001", # move import into type-checking block
    "TC002", # move third-party import into type-checking block
    "TC003", # move stdlib import into type-checking block
    "E741",  # ambiguous variable name
]

[tool.ruff.lint.isort]
known-first-party = ["superseded"]

[tool.ruff.format]
quote-style = "double"